Added UUID functions, integer hash functions, various fixes, and utility rebuild/suinstall targets

+++ include/fdb/dtypeio.h
@@ -85,7 +85,7 @@ FD_FASTOP unsigned int fd_flip_word(unsigned int _w)
 { return ((((_w) << 24) & 0xff000000) | (((_w) << 8) & 0x00ff0000) | 
           (((_w) >> 8) & 0x0000ff00) | ((_w) >>24) );}
 
-FD_FASTOP unsigned int fd_flip_word8(unsigned long long _w)
+FD_FASTOP unsigned long long fd_flip_word8(unsigned long long _w)
 { return (((_w&(0xFF)) << 56) |
 	  ((_w&(0xFF00)) << 48) |
 	  ((_w&(0xFF0000)) << 24) |
